Updated test and post to polls
