--- tsDist/jest.config.js
@@ -1,18 +1,0 @@
-"use strict";
-module.exports = {
-    "roots": [
-        "<rootDir>/src"
-    ],
-    "transform": {
-        "^.+\\.tsx?$": "ts-jest"
-    },
-    "testRegex": "(/__tests__/.*|(\\.|/)(test|spec))\\.tsx?$",
-    "moduleFileExtensions": [
-        "ts",
-        "tsx",
-        "js",
-        "jsx",
-        "json",
-        "node"
-    ],
-};
