--- tsDist/src/server/models/database.js
@@ -1,5 +1,0 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-const loki = require("lokijs");
-const db = new loki("polling-app.db");
-exports.default = db;
