--- tsDist/src/server/server.js
@@ -1,5 +1,0 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-const app_1 = require("./app");
-const port = process.env.PORT || 8000;
-app_1.default.listen(port, () => console.log(`App is running on port: ${port}`));
